+++ packages/core/src/config/logger.ts
@@ -1,0 +1,3 @@
+import { getLogger } from "@midl-xyz/logger";
+
+export const logger = getLogger(["@midl-xyz/midl-js-core"]);