+++ packages/executor/src/config/index.ts
@@ -1,3 +1,4 @@
 export * from "./addresses";
 export * from "./chains";
 export * from "./gasPrice";
+export * from "./logger";