+++ packages/executor-react/CHANGELOG.md
@@ -1,5 +1,14 @@
 # @midl-xyz/midl-js-executor-react
 
+## 2.0.0-next.65
+
+### Patch Changes
+
+- Updated dependencies
+  - @midl-xyz/midl-js-executor@2.0.0-next.65
+  - @midl-xyz/midl-js-core@2.0.0-next.65
+  - @midl-xyz/midl-js-react@2.0.0-next.65
+
 ## 2.0.0-next.64
 
 ### Patch Changes