+++ packages/executor/CHANGELOG.md
@@ -1,5 +1,13 @@
 # @midl-xyz/midl-js-executor
 
+## 2.0.0-next.65
+
+### Patch Changes
+
+- fix(executor): withdrawals overlap deposits
+- Updated dependencies
+  - @midl-xyz/midl-js-core@2.0.0-next.65
+
 ## 2.0.0-next.64
 
 ### Patch Changes